% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{damselfly}
\alias{damselfly}
\title{damselfly data}
\format{
A data frame with 77 observations and 3 variables.
\describe{
  \item{length}{Measured body length of damselflies}
  \item{weight}{Measured weight of damselflies}
  \item{group}{Male age groups; \code{adult} or \code{juvenile} used for body size and weight measurements}
}
}
\source{
Khan, M. K., & Herberstein, M. E. (2021). Male–male interactions select for conspicuous male coloration in damselflies. Animal Behaviour, 176, 157-166.
}
\usage{
damselfly
}
\description{
Data of the body size and weight difference between adult and juvenile males of \emph{Xanthagrion erythroneurum} damselflies.
}
\keyword{datasets}
