% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Durga-package.R
\docType{package}
\name{Durga-package}
\alias{Durga}
\alias{Durga-package}
\title{Durga: Effect Size Estimation and Visualisation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An easy-to-use yet powerful system for plotting grouped data effect sizes. Various types of effect size can be estimated, then plotted together with a representation of the original data. Select from many possible data representations (box plots, violin plots, raw data points etc.), and combine as desired. 'Durga' plots are implemented in base R, so are compatible with base R methods for combining plots, such as 'layout()'. See Khan & McLean (2023) \doi{10.1101/2023.02.06.526960}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KhanKawsar/EstimationPlot}
}

}
\author{
\strong{Maintainer}: Jim McLean \email{jim_mclean@optusnet.com.au} (\href{https://orcid.org/0000-0001-6229-7063}{ORCID})

Authors:
\itemize{
  \item Kawsar Khan \email{bmbkawsar@gmail.com} (\href{https://orcid.org/0000-0003-1795-1315}{ORCID})
}

}
\keyword{internal}
