% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brackets.R
\name{DurgaBrackets}
\alias{DurgaBrackets}
\title{Annotate a \code{DurgaPlot} with confidence brackets}
\usage{
DurgaBrackets(
  plotStats,
  diffs = plotStats$es$group.differences,
  labels = "CI",
  shorten = 1.5,
  tipLength = 2,
  snapTo = 0,
  dataGap = 2.5,
  verticalGap = 1.3,
  textPad = 1.5,
  br.lwd = 1,
  br.col = 1,
  br.lty = 1,
  lb.cex = 1,
  lb.col = 1,
  lb.font = 1,
  ...
)
}
\arguments{
\item{plotStats}{Object returned by the call to \code{\link{DurgaPlot}}}

\item{diffs}{Set of brackets to be displayed as a list of
\code{DurgaGroupDiff} objects}

\item{labels}{Text to display above each bracket. May be NULL, otherwise one
of: \code{"diff"}, \code{"CI"}, \code{"level CI"} or \code{"diff CI"};
a vector of texts to display for each element of \code{diffs}, or a
function called with one argument; a \code{DurgaGroupDiff} object}

\item{shorten}{Amount (mm) to shrink brackets at each end}

\item{tipLength}{Length of bracket tips (mm). May be a vector with length 2;
length of tip at groups 1 and 2 respectively}

\item{snapTo}{Shifts the base of the lowest brackets to a multiple of this
value (mm). Used to improve aesthetics of vertical alignment.}

\item{dataGap}{Vertical distance (mm) between top-most data point and bottom
of bracket}

\item{verticalGap}{Vertical distance (mm) between overlapping brackets}

\item{textPad}{Gap (mm) between bracket and text}

\item{br.lwd, br.col, br.lty}{Graphical parameters that control the bracket
appearance - passed to \code{\link[graphics]{segments}}. May be a single
value or a vector with one value per bracket.}

\item{lb.cex, lb.col, lb.font}{Graphical parameters that control the label
appearance - passed to \code{\link[graphics]{text}}. May be a single
value or a vector with one value per bracket.}

\item{...}{Additional arguments passed to \code{\link[graphics]{text}}}
}
\description{
Brackets are added to a \code{DurgaPlot} that already exists. That means you
must ensure there is sufficient space for the brackets above the plot. To do
this, either specify \code{ylim} to \code{\link{DurgaPlot}}, or create a
large top margin (\code{par(mar = c(...))}) and the turn off the plot frame
(\code{DurgaPlot(..., frame.plot = FALSE)}). In either case, experiment with
the values until the result is visually pleasing. The annotation can be drawn
into the margin without being cropped.
}
\examples{

d <- DurgaDiff(petunia, 1, 2)
# Don't draw frame because brackets will appear in the upper margin
p <- DurgaPlot(d, ef.size = FALSE, frame.plot = FALSE)
DurgaBrackets(p, lb.cex = 0.8, br.lwd = 2, snapTo = 1)

}
