% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{DurgaDiff}
\alias{DurgaDiff}
\title{Estimate group mean differences}
\usage{
DurgaDiff(
  data,
  data.col,
  group.col,
  id.col,
  groups = sort(unique(data[[group.col]])),
  contrasts = "*",
  effect.type = c("unstandardised", "cohens", "hedges"),
  R = 1000,
  boot.params = list(),
  ci.conf = 0.95,
  boot.ci.params = list(),
  na.rm = FALSE
)
}
\arguments{
\item{data}{A data frame containing values to be compared.}

\item{data.col}{Name or index of the column within \code{data} containing the
measurement data.}

\item{group.col}{Name or index of the column within \code{data} containing
the values to group by.}

\item{id.col}{Specify for paired data/repeated measures only. Name or index
of ID column for repeated measures/paired data. Observations for the same
individual must have the same ID. For non-paired data, do not specify an
\code{id.col}.}

\item{groups}{Vector of group names. Defaults to all groups in \code{data} in
\emph{natural} order. If \code{groups} is a named vector, the names are
used as group labels for plotting or printing.}

\item{contrasts}{Specify the pairs of groups to be compared. By default, all
pairwise differences are generated. May be a single string, a vector of
strings, or a matrix. See Details for more information.}

\item{effect.type}{Type of group difference to be estimated Possible types
are: \code{unstandardised}, difference in group means; \code{cohens},
Cohen's d; \code{hedges}, Hedges' g.}

\item{R}{The number of bootstrap replicates.}

\item{boot.params}{Optional list of additional names parameters to pass to
the \code{\link[boot]{boot}} function.}

\item{ci.conf}{Numeric confidence level of the required confidence interval.
Applies to both CI of differences between group means and CI of group
means.}

\item{boot.ci.params}{Optional list of additional names parameters to pass to
the \code{\link[boot]{boot.ci}} function.}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA
values should be stripped before the computation proceeds. If \code{TRUE}
for "paired" data (i.e. \code{id.col} is specified), all rows
(observations) for IDs with missing data are stripped.}
}
\value{
A \code{DurgaDiff} object, which is a list containing:

  \item{\code{group.statistics}}{Matrix with a row for each group, columns
  are: \code{mean}, \code{median}, \code{sd} (standard deviation), \code{se}
  (standard error of the mean), \code{CI.lower} and \code{CI.upper} (lower
  and upper confidence intervals of the mean, confidence level as set by the
  \code{ci.conf} parameter.)}

  \item{\code{group.differences}}{List of \code{DurgaGroupDiff} objects,
  which are \code{boot} objects with added confidence interval information.
  See \code{\link[boot]{boot}} and \code{\link[boot]{boot.ci}}}
  \item{\code{groups}}{Vector of group names}
  \item{\code{group.names}}{Labels used to identify groups}
  \item{\code{effect.type}}{Value of \code{effect.type} parameter}
  \item{\code{effect.name}}{Pretty version of \code{effect.type}}
  \item{\code{data.col}}{Value of \code{data.col} parameter; may be an index
  or a name} \item{\code{data.col.name}}{Name of the \code{data.col} column}
  \item{\code{group.col}}{Value of \code{group.col} parameter; may be an
  index or a name} \item{\code{group.col.name}}{Name of the \code{group.col}
  column} \item{\code{id.col}}{Value of \code{id.col} parameter. May be
  \code{NULL}} \item{\code{paired.data}}{\code{TRUE} if paired differences
  were estimated} \item{\code{data}}{The input data frame}
  \item{\code{call}}{How this function was called}

  A \code{DurgaGroupDiff} object is a \code{boot} object (as returned by
  \code{\link[boot]{boot}}) with added \code{bootci} components (as returned
  by \code{\link[boot]{boot.ci}}) and components identifying the groups used
  to estimate the difference. Particularly relevant members are:

  \item{\code{t0}}{The observed value of the statistic}
  \item{\code{bac[4]}}{The lower endpoint of the confidence interval}
  \item{\code{bac[5]}}{The upper endpoint of the confidence interval}
  \item{\code{groups}}{The difference is estimated on \code{groups[1]} -
  \code{groups[2]}}
}
\description{
Estimates differences between groups in preparation for plotting by
\code{\link{DurgaPlot}}.
}
\details{
The data set must be in \emph{long format}: one column (\code{data.col})
contains the measurement or value to be compared, and another column
(\code{group.col}) the group identity. For repeated measures, a subject
identity column (\code{id.col}) is also required.

The pairs of groups to be compared are defined by the parameter
\code{contrasts}. An asterisk (\code{"*"}, the default) creates contrasts for
all possible pairs of groups. A single string has a format such as
\code{"group1 - group2, group3 - group4"}. A single string such as \code{".-
control"} compares all groups against the \code{"control"} group, i.e. the
\code{"."} expands to all groups except the named group. A vector of strings
looks like \code{c("group1 - group2", "group3 - group4")}. If a matrix is
specified, it must have a column for each contrast, with the first group in
row 1 and the second in row 2.

The formulae for Cohen's d and Hedges' g are from Lakens (2013), equations 1
and 4 respectively. The Cohen's d we use is labelled
\emph{\out{d<sub>s</sub>}} by Lakens (2013). Hedges' g is a corrected version
of Cohen's d, and is more suitable for small sample sizes. For paired (i.e.
repeated measures) Cohen's d, we apply equation 6 (Lakens 2013). For paired
Hedges' g, we apply Hedges' correction to the paired Cohen's d.

Confidence intervals for the estimate are determined using bootstrap
resampling, using the adjusted bootstrap percentile (BCa) method (see
\code{\link[boot]{boot}} and \code{\link[boot]{boot.ci}}). Additional
arguments can be passed to the \code{\link[boot]{boot}}
(\code{\link[boot]{boot.ci}}) by passing a named list of values as the
argument \code{boot.params} (\code{boot.ci.params}).
}
\references{
Lakens, D. (2013). Calculating and reporting effect sizes to facilitate
cumulative science: a practical primer for t-tests and ANOVAs. Frontiers in
Psychology, 4. doi:10.3389/fpsyg.2013.00863
}
\seealso{
\code{\link[boot]{boot}}, \code{\link[boot]{boot.ci}},
  \code{\link{DurgaPlot}}
}
