% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{DurgaTransparent}
\alias{DurgaTransparent}
\title{Returns a transparent version of the specified colour(s).}
\usage{
DurgaTransparent(colour, alpha)
}
\arguments{
\item{colour}{The R colour (or colours) to be made transparent. May be a
colour name, a hexadecimal string such as \code{"#ffbc48"} or a positive
integer.}

\item{alpha}{Transparency, from \code{0} meaning fully opaque (\code{colour}
is returned unchanged), through to \code{1} which is completely transparent
(i.e. invisible).}
}
\value{
A colour or colours that are transparent versions of \code{colour}.
}
\description{
Returns a transparent version of the specified colour(s).
}
\seealso{
[grDevices]{col2rgb}, [grDevices]{rgb}
}
