% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertilisation}
\alias{fertilisation}
\title{Fertilisation data TODO}
\format{
An object of class \code{data.frame} with 30 rows and 3 columns.
}
\usage{
fertilisation
}
\description{
Fertilisation data TODO
}
\keyword{datasets}
