% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insulin}
\alias{insulin}
\title{Insulin data}
\format{
A data frame with 30 observations and 4 variables:
\describe{
  \item{sugar}{Measured blood sugar level}
  \item{time}{Time of blood sugar measurements; \code{before} or \code{after} administering insulin}
  \item{id}{Identifier of indivdual being measured (not in original data set)}
}
}
\source{
TODO reference
}
\usage{
insulin
}
\description{
Results of experiments testing whether insulin reduces blood sugar.
}
\keyword{datasets}
