% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{petunia}
\alias{petunia}
\title{Petunia data}
\format{
A data frame with 64 observations and 3 variables. Plants (self fertilised, inter-cross and westerham-cross) were measured to the tops of their stems when coming into flower. Data collated from Darwin (1877)
\describe{
  \item{height}{Measured height of plants}
  \item{group}{Plant fertilisation groups; \code{westerham_cross}, \code{inter_cross} or \code{self_fertilised}}
  \item{pot_no}{Identifier of pot number in which each plant was grown}
}
}
\source{
Darwin, C. (1877). The effects of cross and self fertilisation in the vegetable kingdom. John Murray, Albemarle Street, London.
}
\usage{
petunia
}
\description{
Charles Darwin's experimental results on petunia plants to determine the difference of plant length between self fertilised and cross-fertilised, either with the same stock (inter-cross) or with a fresh stock (westerham-cross)
}
\seealso{
\link{insulin}, \link{damselfly}
}
\keyword{datasets}
