% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{print.DurgaDiff}
\alias{print.DurgaDiff}
\title{Print a summary of a Durga Difference object}
\usage{
\method{print}{DurgaDiff}(x, ...)
}
\arguments{
\item{x}{An object of class \code{DurgaDiff}.}

\item{...}{Ignored}
}
\description{
This is a method for the function \code{print()} for objects of class
\code{DurgaDiff} created by a call to {DurgaDiff}.
}
\seealso{
\code{\link{print.DurgaPWDiff}}
}
