% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{print.DurgaPWDiff}
\alias{print.DurgaPWDiff}
\title{Print a summary of a Durga group difference object}
\usage{
\method{print}{DurgaPWDiff}(x, ...)
}
\arguments{
\item{x}{An object of class \code{DurgaPWDiff}.}

\item{...}{Ignored}
}
\description{
This is a method for the function \code{print()} for objects of class
\code{DurgaPWDiff}, which is a row in the \code{group.differences} matrix
belonging to an object returned by a call to \code{\link{DurgaDiff}}.
}
\seealso{
\code{\link{DurgaDiff}}, \code{\link{print.DurgaDiff}}
}
