% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{print.SAKDiff}
\alias{print.SAKDiff}
\title{Print a summary of a SAK Difference object}
\usage{
\method{print}{SAKDiff}(x, ...)
}
\arguments{
\item{x}{An object of class \link{\code{SAKDifference}}.}

\item{...}{Ignored}
}
\description{
This is a method for the function \code{print()} for objects of class
"\code{SAKDiff}" created by a call to \link{\code{SAKDifference}}.
}
