% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{print.SAKPWDiff}
\alias{print.SAKPWDiff}
\title{Print a summary of a SAK group difference object}
\usage{
\method{print}{SAKPWDiff}(x, ...)
}
\arguments{
\item{x}{An object of class \link{\code{SAKPWDiff}}.}

\item{...}{Ignored}
}
\description{
This is a method for the function \code{print()} for objects of class
"\code{SAKPWDiff}", which is a row in the \code{group.differences} matrix
belonging to an object returned by a call to \link{\code{SAKDifference}}.
}
