% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{SAKPlot}
\alias{SAKPlot}
\title{Plot grouped data and effect size in base R.}
\usage{
SAKPlot(
  es,
  points = transparent(RColorBrewer::brewer.pal(max(3, length(es$groups)), "Set2"),
    0.4),
  violin = c("left-half", "right-half", "full"),
  violin_border = RColorBrewer::brewer.pal(max(3, length(es$groups)), "Set2"),
  violin_fill = transparent(RColorBrewer::brewer.pal(max(3, length(es$groups)),
    "Set2"), 0.6),
  violin_adj = 1.5,
  violin_width = 0.35,
  violin_trunc_at = 0.05,
  box = "black",
  box_fill = "lightgrey",
  bar = RColorBrewer::brewer.pal(max(3, length(es$groups)), "Set2"),
  bar_fill = transparent(RColorBrewer::brewer.pal(max(3, length(es$groups)), "Set2"),
    0.8),
  ef_size = c(TRUE, "right", "below"),
  ef_size_density = TRUE,
  paired = es$effect.type == "paired",
  central_tendency = c("mean", "median"),
  error_bars = c("CI", "SD", "SE"),
  mean = "grey20",
  mar = c(5, 4, 4, 4) + 0.1,
  xlab = "",
  left_ylab = es$data.col.name,
  right_ylab = "",
  bottom_ylab = "",
  col = c("col1", "col2", "col3"),
  opacity = 0.6,
  points_col = c("col1", "col2", "col3"),
  points_opacity = 0.4,
  las = 1,
  ...
)
}
\arguments{
\item{points}{Vector of colours used to draw data points, one colour per
group. If FALSE or NA, points are not drawn}

\item{violin}{What type of violin plot to display. If FALSE, violin plot is
not drawn}

\item{box}{Colour of boxplot. If FALSE or NA, boxplot is not drawn}

\item{box_fill}{Colour used to fill the bodies of the boxplot. If FALSE or
NA, bodies are not filled}

\item{adj}{Value used to control violin plot smoothness.}
}
\description{

}
