% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insulin}
\alias{insulin}
\title{Insulin data}
\format{
A data frame with 30 observations and 4 variables:
\describe{
  \item{sugar}{}
  \item{time}{}
  \item{id}{}
  \item{experiment}{}
}
}
\source{
TODO
}
\usage{
insulin
}
\description{
Insulin data
}
\keyword{datasets}
