% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{SAKDifference}
\alias{SAKDifference}
\title{Generate differences}
\usage{
SAKDifference(
  data,
  data.col,
  group.col,
  block.col = NULL,
  id.col,
  groups = sort(unique(data[[group.col]])),
  contrast = c("group 1 - group 2"),
  effect.type = c("unstandardised", "cohens", "hedges", "paired"),
  R = 1000,
  ci.type = "bca",
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame...}

\item{data.col}{Name or index of the column within \code{data} containing the measurement data.}

\item{group.col}{Name or index of the column within \code{data} containing the values to group by.}

\item{groups}{Vector of group names. Defaults to all groups in \code{data} in \emph{natural} order.}

\item{effect.type}{Type of difference}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA
values should be stripped before the computation proceeds. If NA values are
stripped and `effect.type` is "paired", all rows (observations) for IDs
with missing data are stripped.}
}
\value{
List containing: bootstrapped mean difference + confidence interval
  of mean difference + individual bootstrapped difference + CI of group 1 +
  CI of group 2 + mean of group 1 + mean of group 2 + median of group 1 +
  median of group 2 + raw data
}
\description{
This is details
}
