% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{SAKDifference}
\alias{SAKDifference}
\title{Calculate group mean differences}
\usage{
SAKDifference(
  data,
  data.col,
  group.col,
  id.col,
  groups = sort(unique(data[[group.col]])),
  contrasts = "*",
  effect.type = c("unstandardised", "cohens", "hedges"),
  R = 1000,
  boot.params = list(),
  ci.conf = 0.95,
  ci.type = "bca",
  na.rm = FALSE
)
}
\arguments{
\item{data}{A data frame containing values to be compared.}

\item{data.col}{Name or index of the column within \code{data} containing the
measurement data.}

\item{group.col}{Name or index of the column within \code{data} containing
the values to group by.}

\item{id.col}{Specify for paired data/repeat measures only. Name or index of
ID column for repeated measures/paired data. For non-paired data, do not
specify an \code{id.col}.}

\item{groups}{Vector of group names. Defaults to all groups in \code{data} in
\emph{natural} order. If \code{groups} is a named vector, the names are
used to identify groups for printing or plotting.}

\item{contrasts}{Specify the pairs of groups to be compared. By default, all
pairwise differences are generated. May be a single string, a vector of
strings, or a matrix. A single string has a format such as \code{"group1 -
group2, group3 - group4"}. A single asterisk, \code{"*"} creates contrasts
for all possible pairs of groups. A single string such as \code{".-
control"} compares all groups against the \code{"control"} group, i.e. the
\code{"."} expands to all groups except the named group. A vector of
strings looks like \code{c("group1 - group2", "group3 - group4")}. If a
matrix is specified, it must have a column for each contrast, with the
first group in row 1 and the second in row 2.}

\item{effect.type}{Type of group difference to be calculated. Possible types
are: \code{unstandardised}, difference in group means; \code{cohens},
Cohen's d; \code{hedges}, Hedges' g.}

\item{R}{The number of bootstrap replicates.}

\item{boot.params}{Optional list of additional names parameters to pass to
the [boot]{boot} function.}

\item{ci.conf}{Numeric confidence level of the required confidence interval.
Applies to both CI of differences between group means and CI of group means.}

\item{ci.type}{A single character  string representing the type of bootstrap
interval required. See the \code{type} parameter to the [boot]{boot.ci}
function for details.}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA
values should be stripped before the computation proceeds. If \code{TRUE}
for "paired" data (i.e. \code{id.col} is specified), all rows
(observations) for IDs with missing data are stripped.}

\item{...}{Any additional parameters are passed to \code{\link[boot]{boot}}.}
}
\value{
List containing:

  \item{\code{groups}}{Vector of group names}
  \item{\code{group.statistics}}{Matrix with a row for each group, columns
  are group mean, median, standard deviation, standard error of the mean,
  lower and upper 95\% confidence intervals of the mean}
  \item{\code{group.differences}}{List of \code{SAKPWDiff} objects, which are
  \code{boot} objects with added confidence interval information. See
  \code{\link[boot]{boot}} and \code{\link[boot]{boot.ci}}}
  \item{\code{effect.type}}{Value of \code{effect.type} parameter}
  \item{\code{effect.name}}{Pretty version of \code{effect.type}}
  \item{\code{data.col}}{Value of \code{data.col} parameter}
  \item{\code{data.col.name}}{Name of the \code{data.col} column}
  \item{\code{group.col}}{Value of \code{group.col} parameter}
  \item{\code{group.col.name}}{Name of the \code{group.col} column}
  \item{\code{data}}{The input data frame} \item{\code{call}}{how this
  function was called} \item{\code{groups}}{Value of \code{groups} parameter}
  \item{\code{group.names}}{Labels used to identify groups}
}
\description{
Calculates differences between groups ready for printing or plotting by
\code{\link{SAKPlot}}.
}
\details{
Data format - long format, one column with measurement (\code{data.col}),
another column with group identity (\code{group.col}). For repeated measures,
a subject identity column is also required (\code{id.col}).

The formulae for Cohen's d and Hedges' g are from Lakens (2013), equations 1
and 4 respectively. The Cohen's d we use is labelled
\emph{\out{d<sub>s</sub>}} by Lakens (2013). Hedges' g is a corrected version
of Cohen's d, and is more suitable for small sample sizes. For paired (i.e.
repeated measures) Cohen's d, we apply equation 6 (Lakens 2013). For paired
Hedges' g, we simply apply Hedges' correction to the paired Cohen's d.
}
\references{
Lakens, D. (2013). Calculating and reporting effect sizes to facilitate
cumulative science: a practical primer for t-tests and ANOVAs. Frontiers in
Psychology, 4. doi:10.3389/fpsyg.2013.00863
}
\seealso{
\code{\link[boot]{boot}}, \code{\link[boot]{boot.ci}},
  \code{\link{SAKPlot}}, \code{\link{print.SAKDiff}}
}
