% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{SAKPlot}
\alias{SAKPlot}
\title{Group and effect size plotting in base R.}
\usage{
SAKPlot(
  es,
  points = TRUE,
  points.method = c("quasirandom", "pseudorandom", "smiley", "maxout", "frowney",
    "minout", "tukey", "tukeyDense", "jitter", "overplot"),
  pch = 19,
  points.params = list(),
  violin = isFALSE(box) && isFALSE(bar),
  violin.shape = c("left-half", "right-half", "full"),
  violin.border = TRUE,
  violin.fill = TRUE,
  violin.adj = 1.5,
  violin.width = 0.35,
  violin.trunc = TRUE,
  box = FALSE,
  box.fill = TRUE,
  box.notch = FALSE,
  box.outline = TRUE,
  box.pars = list(boxwex = 0.8, staplewex = 0.5, outwex = 0.5),
  bar = FALSE,
  bar.fill = TRUE,
  ef.size = c(TRUE, "right", "below"),
  ef.size.violin = TRUE,
  paired = es$effect.type == "paired",
  central.tendency = c("mean", "median"),
  error.bars = c("CI", "SD", "SE"),
  mean = "grey20",
  xlab = "",
  left.ylab = es$data.col.name,
  right.ylab = "",
  ...
)
}
\arguments{
\item{es}{Data returned from a call to \code{\link{SAKDifference}}}

\item{points}{Colour of individual points. If FALSE, points are not plotted,
TRUE plots applies a default colour. May be a vector of colours. If length
1, all points are drawn with the specified colour. If length is the number
of groups, points in each group are drawn with the appropriate colour.
Otherwise, \code{points} should be a vector of colours with a value for
each data point.}

\item{points.method}{Method used to avoid overplotting points. Use
\code{"overplot"} to overplot points and \code{"jitter"} to add random
noise to each x-value. See \code{\link{vipor::offsetX}} for remaining
methods.}

\item{pch}{Symbol to use for plotting points.}

\item{points.params}{List of named parameters to pass on to
\code{\link{graphics::points}}, e.g. \code{SAKPlot(es, points = 1, pch =
21, points.params = list(bg = as.numeric(factor(data$Sex)) + 1))}.}

\item{violin}{What type of violin plot to display. If FALSE, violin plot is
not drawn.}

\item{violin.shape}{Desired violin shape - left-half only (\code{"left"}),
right-half only (\code{"right"}), or a full violin (\code{"full"}).}

\item{violin.border}{Colours of violin borders.}

\item{violin.fill}{Colour used to fill violins.}

\item{violin.adj}{Value used to control violin plot smoothness by adjusting
the kernel density bandwidth. Higher values produce a smoother plot.}

\item{violin.width}{Width of maximum violin horizontal extents, as a
proportion of the distance between groups.}

\item{violin.trunc}{Numeric value that specifies what vertical proportion of
the violin is truncated.}

\item{box}{If not FALSE, draw a box-and-whisker plot of the grouped values.
Value may be a colour, in which case the box borders are plotted with the
colour(s). See \code{\link{graphics::boxplot}}.}

\item{box.fill}{Colour used to fill the bodies of the box-and-whisker plot.
If FALSE or NA, bodies are not filled}

\item{box.notch}{If TRUE, draws notches in the sides of the boxes. See
\code{\link{graphics::boxplot.stats}} for the calculations used.}

\item{box.pars}{List with additional graphical parameters to control the box
plot. See \code{\link{graphics::bxp}} graphical parameters for a complete
list.}

\item{bar}{If not FALSE, draw a bar plot of the group means.}

\item{bar.fill}{Colour used to fill bars.}

\item{ef.size}{If not FALSE, effect sizes are plotted. Effect sizes are
plotted to the right of the main plot if there is only one effect size to
plot and \code{ef.size != "below"}.}

\item{ef.size.violin}{If not FALSE, boostrapped effect size estimates are
show as a violin plot.}
}
\value{
\code{es} invisibly.
}
\description{
Plot grouped data and effect size in base R, with control over a large range
of possible display formats and options. To plot your data, first calculate
group differences by calling \code{\link{SAKDifference}}, then pass the
result to \code{\link{SAKPlot}}.
}
\details{
Group data may be visualised in multiple ways: \code{points}, \code{violin},
\code{box} and \code{bar}. Each visualisation type is controlled by a set of
parameters. To display a type, for example box plots, specify \code{box =
TRUE}. Rather than \code{TRUE}, you may specify a colour, which is used is
the border/outline for the boxes. You may also specify a vector of colours,
one for each group. For \code{points}, you may specify a colour for each
individual point. When colours are not specified, they are selected from an
\code{\link{RColorBrewer}} qualitative palette.
}
\seealso{
\code{\link{SAKDifference}}, \code{\link{vipor::offsetX}},
  \code{\link{graphics::boxplot}}, \code{\link{graphics::bxp}}
}
