% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{SAKPlot}
\alias{SAKPlot}
\title{Group and effect size plotting in base R.}
\usage{
SAKPlot(
  es,
  contrasts,
  group.dx = 0,
  points = TRUE,
  points.method = c("quasirandom", "pseudorandom", "smiley", "maxout", "frowney",
    "minout", "tukey", "tukeyDense", "jitter", "overplot"),
  points.dx = group.dx,
  points.params = list(),
  violin = isFALSE(box) && isFALSE(bar),
  violin.shape = c("left-half", "right-half", "full"),
  violin.fill = TRUE,
  violin.adj = 1.5,
  violin.width = 0.35,
  violin.trunc = TRUE,
  violin.dx = group.dx,
  box = FALSE,
  box.fill = TRUE,
  box.outline = TRUE,
  box.notch = FALSE,
  box.pars = list(boxwex = 0.8, staplewex = 0.5, outwex = 0.5),
  box.dx = group.dx,
  bar = FALSE,
  bar.fill = TRUE,
  bar.width = 0.8,
  bar.dx = group.dx,
  ef.size = TRUE,
  ef.size.position = c("right", "below"),
  ef.size.violin = TRUE,
  ef.size.violin.shape = c("right-half", "left-half", "full"),
  ef.size.pch = 17,
  ef.size.dx = group.dx,
  ef.size.ticks = NULL,
  ef.size.las = 0,
  ef.size.label = es$effect.name,
  paired = es$paired.data,
  central.tendency = isFALSE(box),
  central.tendency.type = c("mean", "median"),
  central.tendency.symbol = c("point", "segment"),
  central.tendency.params = list(),
  central.tendency.dx = group.dx,
  error.bars = central.tendency,
  error.bars.type = c("CI", "SD", "SE"),
  axis.dx = group.dx,
  left.ylab = es$data.col.name,
  left.las = 0,
  ...
)
}
\arguments{
\item{es}{Data returned from a call to \code{\link{SAKDifference}}}

\item{contrasts}{Set of contrasts (i.e. group comparisons) to be plotted. By
default, plots the contrasts specified in the call to
\code{\link{SAKDifference}}. If contrasts were not explicitly specified,
plots 2nd and subsequent groups minus the first group. May be a single
string, a vector of strings, or a matrix. A single string has a format such
as \code{"group1 - group2, group3 - group4"}. A single asterisk, \code{"*"}
creates contrasts for all possible pairs of groups. A single string such as
\code{".- control"} compares all groups against the \code{"control"} group,
i.e. the \code{"."} expands to all groups except the named group. A vector
of strings looks like \code{c("group1 - group2", "group3 - group4")}. If a
matrix is specified, it must have a column for each contrast, with the
first group in row 1 and the second in row 2. See also the \code{contrasts}
parameter to \code{\link{SAKDifference}}. It is an error to attempt to plot
a contrast that was not estimated by \code{\link{SAKDifference}}.}

\item{group.dx}{Used to shift group centres horizontally. E.g.,
\code{group.dx = c(0.1, -0.1)} will group into pairs. Individual components
can be shifted independently using the appropriate \code{*.dx} parameters.}

\item{points}{If not FALSE, points are plotted. If \code{TRUE}, points are
displayed with a default colour. You may specify a vector of colours; if
length 1, all points are drawn with the specified colour. If length is less
than the number of data points, points in each group are drawn with the
appropriate colour (extra colours are ignored). Otherwise, \code{points}
should be a vector of colours with a value for each data point.}

\item{points.method}{Method used to avoid overplotting points. Use
\code{"overplot"} to overplot points and \code{"jitter"} to add random
noise to each x-value. See \code{\link[vipor]{offsetX}} for remaining
methods.}

\item{points.dx}{Horizontal shift to be applied to points in each group.}

\item{points.params}{List of named parameters to pass on to
\code{\link[graphics]{points}}, e.g. \code{SAKPlot(es, points = "black",
points.params = list(pch = 21, bg = as.numeric(factor(data$Sex)) + 1))}.}

\item{violin}{If not FALSE, violin plots are drawn. If \code{TRUE}, violins
are drawn in default colours. Otherwise specifies the colour of the violin
borders.}

\item{violin.shape}{Desired violin shape - left-half only (\code{"left"}),
right-half only (\code{"right"}), or a full violin (\code{"full"}).}

\item{violin.fill}{Colour used to fill violins.}

\item{violin.adj}{Value used to control violin plot smoothness by adjusting
the kernel density bandwidth. Higher values produce a smoother plot.}

\item{violin.width}{Width of maximum violin horizontal extents, as a
proportion of the distance between groups.}

\item{violin.trunc}{Numeric value that specifies what vertical proportion of
the violin is truncated.}

\item{violin.dx}{Horizontal shift to be applied to each violin.}

\item{box}{If not FALSE, draw a box-and-whisker plot of the grouped values.
Value may be a colour, in which case the box borders are plotted with the
colour(s). See \code{\link[graphics]{boxplot}}.}

\item{box.fill}{Colour used to fill the bodies of the box-and-whisker plot.
If FALSE or NA, bodies are not filled.}

\item{box.outline}{If FALSE, don't draw outliers with the box plot.}

\item{box.notch}{If TRUE, draws notches in the sides of the boxes. See
\code{\link[graphics]{boxplot.stats}} for the calculations used.}

\item{box.pars}{List with additional graphical parameters to control the box
plot. See \code{\link[graphics]{bxp}} graphical parameters for a complete
list.}

\item{box.dx}{Horizontal shift to be applied to each box.}

\item{bar}{If not FALSE, draw a bar plot of the group means or medians,
according to \code{central.tendency}. May be \code{TRUE} or a colour.}

\item{bar.fill}{Colour used to fill bars.}

\item{bar.width}{Width of bars.}

\item{bar.dx}{Horizontal shift to be applied to each bar.}

\item{ef.size}{If not FALSE, effect sizes are plotted. May be \code{TRUE} or
a colour.}

\item{ef.size.position}{Effect sizes are plotted to the right of the main
plot if there is only one effect size to plot and \code{ef.size.position !=
"below"}. If the effect size is drawn to the right, you will need to
increase the size of the right margin before plotting (see
\code{\link[graphics:par]{par(mar = ...)}}).}

\item{ef.size.violin}{If not FALSE, boostrapped effect size estimates are
show as a violin plot. May be a colour, used for the violin border, and a
transparent version is used for the violin fill.}

\item{ef.size.violin.shape}{Shape of the effect size violin.}

\item{ef.size.pch}{Symbol to represent mean effect size.}

\item{ef.size.dx}{Horizontal shift to be applied to each effect size.}

\item{ef.size.ticks}{Optional locations and labels for ticks on the effect
size y-axis. E.g. to interpret effect size using Cohen's default values,
specif \code{ef.size.ticks = c("Large negative effect" = -0.8, "Medium
negative effect" = -0.5, "Small negative effect" = -0.2, "No effect" = 0,
"Small positive effect" = 0.2, "Medium positive effect" = 0.5, "Large
positive effect" = 0.8)}}

\item{ef.size.las}{Orientation of tick labels on the effect size axis (0 =
parallel to axis, 1 = horizontal).}

\item{ef.size.label}{Label to display on y-axis for effect size.}

\item{paired}{If \code{TRUE}, lines are drawn joining the individual data
points.}

\item{central.tendency}{If not FALSE, a visual indicator of central tendency
is drawn. May be a colour, in which case it is used for mean/median and
error bars.}

\item{central.tendency.type}{Should the indicated measure of central tendency
be \code{"mean"} or \code{"median"}?}

\item{central.tendency.symbol}{Should central tendency be shown as a point or
a horizontal line segment?}

\item{central.tendency.params}{Additional arguments to be passed to
\code{\link[graphics]{points}} (if \code{central.tendency.symbol ==
"point"}) or \code{\link[graphics]{segments}} (if
\code{central.tendency.symbol == "segment"}).}

\item{central.tendency.dx}{Horizontal shift to apply to central tendency
indicator and error bars.}

\item{error.bars}{Should error bars be displayed? May be the colour to be
used for error bars.}

\item{error.bars.type}{Should error bars depict 95%% confidence intervals of
the mean (\code{"CI"}), standard deviation (\code{"SD"}) or standard error
(\code{"SE"})?}

\item{axis.dx}{Horizontal shifts to be applied to each x-axis tick and label.}

\item{left.ylab}{Left-hand y-axis label.}

\item{left.las}{Orientation of axis labels on left-hand y-axis label (0 =
parallel to axis, 1 = horizontal).}

\item{...}{Additional arguments are passed on to the
\code{\link[graphics]{plot}} function.}
}
\value{
\code{es} invisibly.
}
\description{
Plot grouped data and effect size in base R, with control over a large range
of possible display formats and options. To plot your data, first calculate
group differences by calling \code{\link{SAKDifference}}, then pass the
result to \code{\link{SAKPlot}}. Parameters are grouped according to the
component they affect, so all parameters that affect box plots are prefixed
with \code{box}.
}
\details{
Group data may be visualised in multiple ways: \code{points}, \code{violin},
\code{box} and \code{bar}. Each visualisation type is controlled by a set of
parameters with the same prefix. To display a type, for example box plots,
specify \code{box = TRUE}. Rather than \code{box = TRUE}, you may specify a
colour (e.g. \code{box \ "blue"}), which is used as the border/outline for
the boxes. You may also specify a vector of colours, one for each group. For
\code{points}, you may specify a colour for each individual point. When
colours are not specified, they are selected from an
\code{\link{RColorBrewer}} qualitative palette.
}
\references{
Gardner, M. J., & Altman, D. G. (1986). Confidence intervals rather than P
values: estimation rather than hypothesis testing. Br Med J (Clin Res Ed),
292(6522), 746-750. doi:10.1136/bmj.292.6522.746

Cumming, G. (2012). Understanding the new statistics : effect sizes,
confidence intervals, and meta-analysis (1st edition ed.). New York:
Routledge.
}
\seealso{
\code{\link{SAKDifference}}, \code{\link[vipor]{offsetX}},
  \code{\link[graphics]{boxplot}}, \code{\link[graphics]{bxp}}
}
